move SVG to core
